add jquery-ui-rails

+++ app/controllers/admin/events_controller.rb
@@ -49,9 +49,17 @@ class Admin::EventsController < AdminController
      total = 0
      Array(params[:ids]).each do |event_id|
        event = Event.find(event_id)
-       event.destroy
-       total  = 1
-     end
+
+        if params[:commit] == I18n.t(:bulk_update)
+          event.status = params[:event_status]
+          if event.save
+            total += 1
+          end
+        elsif params[:commit] == I18n.t(:bulk_delete)
+          event.destroy
+          total += 1
+        end
+      end
 
      flash[:alert] = "成功完成 #{total} 笔"
      redirect_to admin_events_path
@@ -62,7 +70,10 @@ class Admin::EventsController < AdminController
       @event.row_order_position = params[:position]
       @event.save!
 
-       redirect_to admin_events_path
+       respond_to do |format|
+         format.html {redirect_to admin_events_path}
+         format.json {render :json => { :message => "ok"}}
+       end
     end
